fix: исправлены 2 ошибки компиляции
- SettingsScreen.java:285 - getter лямбда triggerDelay изменена с float на double (DoubleSupplier)
- HitboxExpander.java:85 - убрана попытка использовать Attributes.ATTACK_REACH которого нет в 1.16.5

+++ src/main/java/com/example/cosmetics/gui/SettingsScreen.java
@@ -282,7 +282,7 @@ public class SettingsScreen extends Screen {
         // ---- TriggerBot ------------------------------------------------------
         else if (feature == FeatureType.TRIGGER_BOT) {
             sliders.add(new Slider(sx, sy + i++ * ROW_H, sw, 17, "Delay ms", 0F, 500F,
-                    () -> (float) fs.triggerDelay, v -> fs.triggerDelay = (long) v) {
+                    () -> (double) fs.triggerDelay, v -> fs.triggerDelay = (long)(float) v) {
                 @Override public String formatValue() {
                     return fs.triggerDelay == 0 ? "0 ms (instant)" : fs.triggerDelay + " ms";
                 }
